Github/03_Regular AI meeting minutes/requirements.txt
# AI Meeting Minutes Automation System
# Python 3.10+

# Browser automation for tldv transcript scraping
playwright>=1.40.0

# Anti-detection for Playwright (bypass automation detection)
playwright-stealth>=1.0.0

# Anthropic Claude API (using Haiku for cost efficiency)
anthropic>=0.40.0

# HTTP requests for Teams Workflows and Graph API
requests>=2.31.0

# Microsoft Authentication Library (for OneNote Graph API)
msal>=1.28.0

# Environment variable management
python-dotenv>=1.0.0
